Makefile + libft.h

+++ libft/ft_strnstr.c
@@ -12,10 +12,10 @@
 
 #include <stddef.h>
 
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
 {
-	size_t	i;
-	size_t	j;
+	unsigned int	i;
+	unsigned int	j;
 
 	i = 0;
 	if (!*needle)
